# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bengkol Workshop & Repair-Shop Management API
  description: |
    Production-ready REST API for the Bengkol auto repair-shop platform.
    Connects vehicle owners with nearby repair shops for slot booking, real-time queues, service catalogs, inventory, and reviews.
  version: 1.0.0
  contact:
    name: Bengkol Engineering
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api-dev.bengkol.com
    description: Development (Develop) environment
  - url: https://api-uat.bengkol.com
    description: User Acceptance Testing (UAT) environment
  - url: https://api-staging.bengkol.com
    description: Staging environment
  - url: https://api.bengkol.com
    description: Production (Prod) environment
paths:
...
  /api/v1/workshops:
    get:
      summary: List workshops
      tags:
        - Workshops
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
      responses:
        '200':
          description: List of workshops
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopListResponse'
    post:
      summary: Create a new workshop
      tags:
        - Workshops
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateWorkshopRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkshopRequest'
      responses:
        '201':
          description: Workshop created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopDetailResponse'
  /api/v1/workshops/nearby:
    get:
      summary: Find nearby workshops via PostGIS spatial search
      tags:
        - Workshops
      parameters:
        - in: query
          name: lat
          required: true
          schema:
            type: number
            format: double
        - in: query
          name: lng
          required: true
          schema:
            type: number
            format: double
        - in: query
          name: radius
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Nearby workshops ordered by distance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearbyWorkshopsResponse'
  /api/v1/workshops/{id}:
    get:
      summary: Get workshop detail
      tags:
        - Workshops
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workshop details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopDetailResponse'
    patch:
      summary: Update workshop profile
      tags:
        - Workshops
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateWorkshopRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkshopRequest'
      responses:
        '200':
          description: Updated workshop details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopDetailResponse'
  /api/v1/workshops/{id}/photos/{photoID}:
    get:
      summary: Retrieve workshop photo binary data
      description: Returns the raw binary image stream stored directly in the database.
      tags:
        - Workshops
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: photoID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Binary photo image stream
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workshops/{id}/operating-hours:
    get:
      summary: Get workshop operating hours
      tags:
        - Workshops
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of operating schedules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatingHoursResponse'
    put:
      summary: Set weekly operating hours
      tags:
        - Workshops
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/OperatingHourInput'
      responses:
        '200':
          description: Updated operating hours
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatingHoursResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/APIError'
...
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
          format: int64
        total_pages:
          type: integer
...
    CreateWorkshopRequest:
      type: object
      required:
        - name
        - address
        - latitude
        - longitude
        - phone
        - photos
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        phone:
          type: string
        photos:
          type: array
          minItems: 3
          items:
            type: string
            format: binary
          description: Workshop photos in binary format (at least 3 photos required)
          example:
            - "https://example.com/photos/1.jpg"
            - "https://example.com/photos/2.jpg"
            - "https://example.com/photos/3.jpg"
        employees:
          type: array
          description: Optional initial employees to add simultaneously when creating the workshop
          items:
            $ref: '#/components/schemas/CreateEmployeeInput'
    UpdateWorkshopRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        phone:
          type: string
        photos:
          type: array
          minItems: 3
          items:
            type: string
            format: binary
          description: Workshop photos in binary format (at least 3 photos required)
          example:
            - "https://example.com/photos/1.jpg"
            - "https://example.com/photos/2.jpg"
            - "https://example.com/photos/3.jpg"
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
    OperatingHourInput:
      type: object
      required:
        - day_of_week
      properties:
        day_of_week:
          type: integer
        open_time:
          type: string
        close_time:
          type: string
        is_closed:
          type: boolean
    OperatingHour:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workshop_id:
          type: string
          format: uuid
        day_of_week:
          type: integer
        open_time:
          type: string
        close_time:
          type: string
        is_closed:
          type: boolean
    Workshop:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        phone:
          type: string
        photos:
          type: array
          items:
            type: string
            format: binary
          example:
            - "/api/v1/workshops/3fa85f64-5717-4562-b3fc-2c963f66afa6/photos/9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
            - "/api/v1/workshops/3fa85f64-5717-4562-b3fc-2c963f66afa6/photos/a2b3c4d5-5717-4562-b3fc-2c963f66afa6"
            - "/api/v1/workshops/3fa85f64-5717-4562-b3fc-2c963f66afa6/photos/b3c4d5e6-5717-4562-b3fc-2c963f66afa6"
        rating:
          type: number
          format: float
        review_count:
          type: integer
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
        employees:
          type: array
          description: List of employees belonging to the workshop
          items:
            $ref: '#/components/schemas/WorkshopEmployee'
    WorkshopListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Workshop'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    NearbyWorkshopsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Workshop'
    WorkshopDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Workshop'
    OperatingHoursResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/OperatingHour'
...
    EmployeeRole:
      type: string
      enum:
        - MECHANIC
        - ADMIN_CASHIER
        - ADMIN_INVENTORY
        - ADMIN_BOTH
        - MANAGER
        - OWNER
      description: Specific employee access role within a workshop
    EmployeeStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
    EmployeePermissions:
      type: object
      properties:
        can_manage_employees:
          type: boolean
        can_manage_inventory:
          type: boolean
        can_access_cashier:
          type: boolean
        can_access_repair_jobs:
          type: boolean
        can_manage_workshop_operations:
          type: boolean
    WorkshopEmployee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workshop_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/EmployeeRole'
        status:
          $ref: '#/components/schemas/EmployeeStatus'
        specialization:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        permissions:
          $ref: '#/components/schemas/EmployeePermissions'
    CreateEmployeeInput:
      type: object
      required:
        - name
        - phone
        - role
      properties:
        user_id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/EmployeeRole'
        status:
          $ref: '#/components/schemas/EmployeeStatus'
        specialization:
          type: string
        notes:
          type: string
...